Fix rounding error in conversion to yolo format

+++ albumentations/augmentations/bbox_utils.py
@@ -289,8 +289,8 @@ def convert_bbox_from_albumentations(bbox, target_format, rows, cols, check_vali
     elif target_format == "yolo":
         # https://github.com/pjreddie/darknet/blob/f6d861736038da22c9eb0739dca84003c5a5e275/scripts/voc_label.py#L12
         (x_min, y_min, x_max, y_max), tail = bbox[:4], bbox[4:]
-        x = (x_min + x_max) / 2 - 1
-        y = (y_min + y_max) / 2 - 1
+        x = int((x_min + x_max) / 2 - 1)
+        y = int((y_min + y_max) / 2 - 1)
         width = x_max - x_min
         height = y_max - y_min
         bbox = normalize_bbox((x, y, width, height) + tail, rows, cols)
